Reorganized api clients and student queries and mutations
