+++ web/app/trainers/[id]/page.tsx
@@ -9,9 +9,9 @@ import { useRouter, useParams } from "next/navigation"
 import { useState, useEffect } from "react"
 import Layout from "@/components/layout"
 import TrainerModal from "@/components/trainer-modal"
-import { ListTrainersDto, TrainerResponseDto, update } from "@/lib/api-client"
+import { ListTrainersDto, TrainerResponseDto, update } from "@/lib/custom-api-client"
 import { findTrainerByIdOptions, updateTrainerAndUserMutation } from "@/lib/api-client/@tanstack/react-query.gen"
-import { apiClient } from "@/lib/client"
+import { customApiClient } from "@/lib/custom-api-client"
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
 
 export default function TrainerProfilePage() {
@@ -25,7 +25,7 @@ export default function TrainerProfilePage() {
   const { data: trainerData, isLoading, error } = useQuery({
     ...findTrainerByIdOptions({
       path: { id: params.id as string },
-      client: apiClient,
+      client: customApiClient,
       security: [{
         type: "http",
         scheme: "bearer",
@@ -72,7 +72,7 @@ export default function TrainerProfilePage() {
       await updateTrainer({
         path: { id: String(updatedTrainer?.id) },
         body: updatedTrainer,
-        client: apiClient,
+        client: customApiClient,
         security: [{
           type: "http",
           scheme: "bearer",
