+++ web/components/student-form.tsx
@@ -11,7 +11,7 @@ import { Textarea } from "@/components/ui/textarea"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { Checkbox } from "@/components/ui/checkbox"
 import { User, Plus, Trash2 } from "lucide-react"
-import {AnamnesisResponseDto, PhysicalImpairmentResponseDto} from "@/lib/api-client";
+import {AnamnesisResponseDto, PhysicalImpairmentResponseDto} from "@/lib/custom-api-client";
 import {hasInsomniaTypes, impairmentTypes} from "@/lib/students-data";
 
 interface PhysicalImpairment {
